Upgrade terraform-provider-azurerm to v4.69.0 (#3483)
This PR was generated via `$ upgrade-provider pulumi/pulumi-azure
--repo-path .`.

---

- Upgrading terraform-provider-azurerm from 4.66.0  to 4.69.0.
	Fixes #3478
	Fixes #3474
	Fixes #3469


---

### Fixes applied to unblock upgrade

- Rebased `patches/0011-Update-documentation.patch` onto azurerm
`v4.69.0` and resolved the resulting conflicts in
`upstream/website/docs/d/storage_account_blob_container_sas.html.markdown`,
`upstream/website/docs/d/storage_account_sas.html.markdown`, and
`upstream/website/docs/r/public_ip.html.markdown` by keeping the newer
upstream facts while preserving the Pulumi doc-cleanup intent.
- Added explicit token mappings in `provider/resources.go` for the new
upstream `azurerm_managed_devops_pool` resource/data source and
`azurerm_resource_provider_feature_registration` resource so `make
tfgen` could assign Pulumi tokens successfully.

+++ sdk/python/pulumi_azure/storage/outputs.py
@@ -3369,33 +3369,70 @@ class GetAccountAzureFilesAuthenticationActiveDirectoryResult(dict):
 @pulumi.output_type
 class GetAccountBlobContainerSASPermissionsResult(dict):
     def __init__(__self__, *,
-                 add: _builtins.bool,
-                 create: _builtins.bool,
-                 delete: _builtins.bool,
-                 list: _builtins.bool,
-                 read: _builtins.bool,
-                 write: _builtins.bool):
+                 add: Optional[_builtins.bool] = None,
+                 create: Optional[_builtins.bool] = None,
+                 delete: Optional[_builtins.bool] = None,
+                 delete_version: Optional[_builtins.bool] = None,
+                 execute: Optional[_builtins.bool] = None,
+                 find: Optional[_builtins.bool] = None,
+                 list: Optional[_builtins.bool] = None,
+                 move: Optional[_builtins.bool] = None,
+                 ownership: Optional[_builtins.bool] = None,
+                 permissions: Optional[_builtins.bool] = None,
+                 read: Optional[_builtins.bool] = None,
+                 set_immutability_policy: Optional[_builtins.bool] = None,
+                 tags: Optional[_builtins.bool] = None,
+                 write: Optional[_builtins.bool] = None):
         """
         :param _builtins.bool add: Should Add permissions be enabled for this SAS?
         :param _builtins.bool create: Should Create permissions be enabled for this SAS?
         :param _builtins.bool delete: Should Delete permissions be enabled for this SAS?
+        :param _builtins.bool delete_version: Should Delete version permissions be enabled for this SAS?
+        :param _builtins.bool execute: Should Execute permissions be enabled for this SAS?
+        :param _builtins.bool find: Should Find permissions be enabled for this SAS?
         :param _builtins.bool list: Should List permissions be enabled for this SAS?
-               
-               Refer to the [SAS creation reference from Azure](https://docs.microsoft.com/rest/api/storageservices/create-service-sas)
-               for additional details on the fields above.
+        :param _builtins.bool move: Should Move permissions be enabled for this SAS?
+        :param _builtins.bool ownership: Should Ownership permissions be enabled for this SAS?
+        :param _builtins.bool permissions: Should Permissions permissions be enabled for this SAS?
         :param _builtins.bool read: Should Read permissions be enabled for this SAS?
+        :param _builtins.bool set_immutability_policy: Should Set Immutability Policy permissions be enabled for this SAS?
+        :param _builtins.bool tags: Should Tags permissions be enabled for this SAS?
         :param _builtins.bool write: Should Write permissions be enabled for this SAS?
+               
+               > **Note:** Refer to the [SAS creation reference from Azure](https://docs.microsoft.com/rest/api/storageservices/create-service-sas) for additional details on the fields above.
         """
-        pulumi.set(__self__, "add", add)
-        pulumi.set(__self__, "create", create)
-        pulumi.set(__self__, "delete", delete)
-        pulumi.set(__self__, "list", list)
-        pulumi.set(__self__, "read", read)
-        pulumi.set(__self__, "write", write)
+        if add is not None:
+            pulumi.set(__self__, "add", add)
+        if create is not None:
+            pulumi.set(__self__, "create", create)
+        if delete is not None:
+            pulumi.set(__self__, "delete", delete)
+        if delete_version is not None:
+            pulumi.set(__self__, "delete_version", delete_version)
+        if execute is not None:
+            pulumi.set(__self__, "execute", execute)
+        if find is not None:
+            pulumi.set(__self__, "find", find)
+        if list is not None:
+            pulumi.set(__self__, "list", list)
+        if move is not None:
+            pulumi.set(__self__, "move", move)
+        if ownership is not None:
+            pulumi.set(__self__, "ownership", ownership)
+        if permissions is not None:
+            pulumi.set(__self__, "permissions", permissions)
+        if read is not None:
+            pulumi.set(__self__, "read", read)
+        if set_immutability_policy is not None:
+            pulumi.set(__self__, "set_immutability_policy", set_immutability_policy)
+        if tags is not None:
+            pulumi.set(__self__, "tags", tags)
+        if write is not None:
+            pulumi.set(__self__, "write", write)
 
     @_builtins.property
     @pulumi.getter
-    def add(self) -> _builtins.bool:
+    def add(self) -> Optional[_builtins.bool]:
         """
         Should Add permissions be enabled for this SAS?
         """
@@ -3403,7 +3440,7 @@ class GetAccountBlobContainerSASPermissionsResult(dict):
 
     @_builtins.property
     @pulumi.getter
-    def create(self) -> _builtins.bool:
+    def create(self) -> Optional[_builtins.bool]:
         """
         Should Create permissions be enabled for this SAS?
         """
@@ -3411,36 +3448,99 @@ class GetAccountBlobContainerSASPermissionsResult(dict):
 
     @_builtins.property
     @pulumi.getter
-    def delete(self) -> _builtins.bool:
+    def delete(self) -> Optional[_builtins.bool]:
         """
         Should Delete permissions be enabled for this SAS?
         """
         return pulumi.get(self, "delete")
 
     @_builtins.property
+    @pulumi.getter(name="deleteVersion")
+    def delete_version(self) -> Optional[_builtins.bool]:
+        """
+        Should Delete version permissions be enabled for this SAS?
+        """
+        return pulumi.get(self, "delete_version")
+
+    @_builtins.property
     @pulumi.getter
-    def list(self) -> _builtins.bool:
+    def execute(self) -> Optional[_builtins.bool]:
+        """
+        Should Execute permissions be enabled for this SAS?
+        """
+        return pulumi.get(self, "execute")
+
+    @_builtins.property
+    @pulumi.getter
+    def find(self) -> Optional[_builtins.bool]:
+        """
+        Should Find permissions be enabled for this SAS?
+        """
+        return pulumi.get(self, "find")
+
+    @_builtins.property
+    @pulumi.getter
+    def list(self) -> Optional[_builtins.bool]:
         """
         Should List permissions be enabled for this SAS?
-
-        Refer to the [SAS creation reference from Azure](https://docs.microsoft.com/rest/api/storageservices/create-service-sas)
-        for additional details on the fields above.
         """
         return pulumi.get(self, "list")
 
     @_builtins.property
     @pulumi.getter
-    def read(self) -> _builtins.bool:
+    def move(self) -> Optional[_builtins.bool]:
+        """
+        Should Move permissions be enabled for this SAS?
+        """
+        return pulumi.get(self, "move")
+
+    @_builtins.property
+    @pulumi.getter
+    def ownership(self) -> Optional[_builtins.bool]:
+        """
+        Should Ownership permissions be enabled for this SAS?
+        """
+        return pulumi.get(self, "ownership")
+
+    @_builtins.property
+    @pulumi.getter
+    def permissions(self) -> Optional[_builtins.bool]:
+        """
+        Should Permissions permissions be enabled for this SAS?
+        """
+        return pulumi.get(self, "permissions")
+
+    @_builtins.property
+    @pulumi.getter
+    def read(self) -> Optional[_builtins.bool]:
         """
         Should Read permissions be enabled for this SAS?
         """
         return pulumi.get(self, "read")
 
     @_builtins.property
+    @pulumi.getter(name="setImmutabilityPolicy")
+    def set_immutability_policy(self) -> Optional[_builtins.bool]:
+        """
+        Should Set Immutability Policy permissions be enabled for this SAS?
+        """
+        return pulumi.get(self, "set_immutability_policy")
+
+    @_builtins.property
     @pulumi.getter
-    def write(self) -> _builtins.bool:
+    def tags(self) -> Optional[_builtins.bool]:
+        """
+        Should Tags permissions be enabled for this SAS?
+        """
+        return pulumi.get(self, "tags")
+
+    @_builtins.property
+    @pulumi.getter
+    def write(self) -> Optional[_builtins.bool]:
         """
         Should Write permissions be enabled for this SAS?
+
+        > **Note:** Refer to the [SAS creation reference from Azure](https://docs.microsoft.com/rest/api/storageservices/create-service-sas) for additional details on the fields above.
         """
         return pulumi.get(self, "write")
 
@@ -3517,45 +3617,54 @@ class GetAccountIdentityResult(dict):
 @pulumi.output_type
 class GetAccountSASPermissionsResult(dict):
     def __init__(__self__, *,
-                 add: _builtins.bool,
-                 create: _builtins.bool,
-                 delete: _builtins.bool,
-                 filter: _builtins.bool,
-                 list: _builtins.bool,
-                 process: _builtins.bool,
-                 read: _builtins.bool,
-                 tag: _builtins.bool,
-                 update: _builtins.bool,
-                 write: _builtins.bool):
+                 add: Optional[_builtins.bool] = None,
+                 create: Optional[_builtins.bool] = None,
+                 delete: Optional[_builtins.bool] = None,
+                 filter: Optional[_builtins.bool] = None,
+                 list: Optional[_builtins.bool] = None,
+                 process: Optional[_builtins.bool] = None,
+                 read: Optional[_builtins.bool] = None,
+                 tag: Optional[_builtins.bool] = None,
+                 update: Optional[_builtins.bool] = None,
+                 write: Optional[_builtins.bool] = None):
         """
         :param _builtins.bool add: Should Add permissions be enabled for this SAS?
         :param _builtins.bool create: Should Create permissions be enabled for this SAS?
         :param _builtins.bool delete: Should Delete permissions be enabled for this SAS?
         :param _builtins.bool filter: Should Filter by Index Tags permissions be enabled for this SAS?
-               
-               Refer to the [SAS creation reference from Azure](https://docs.microsoft.com/rest/api/storageservices/constructing-an-account-sas)
-               for additional details on the fields above.
         :param _builtins.bool list: Should List permissions be enabled for this SAS?
         :param _builtins.bool process: Should Process permissions be enabled for this SAS?
         :param _builtins.bool read: Should Read permissions be enabled for this SAS?
         :param _builtins.bool tag: Should Get / Set Index Tags permissions be enabled for this SAS?
         :param _builtins.bool update: Should Update permissions be enabled for this SAS?
         :param _builtins.bool write: Should Write permissions be enabled for this SAS?
+               
+               > **Note:** Refer to the [SAS creation reference from Azure](https://docs.microsoft.com/rest/api/storageservices/constructing-an-account-sas) for additional details on the fields above.
         """
-        pulumi.set(__self__, "add", add)
-        pulumi.set(__self__, "create", create)
-        pulumi.set(__self__, "delete", delete)
-        pulumi.set(__self__, "filter", filter)
-        pulumi.set(__self__, "list", list)
-        pulumi.set(__self__, "process", process)
-        pulumi.set(__self__, "read", read)
-        pulumi.set(__self__, "tag", tag)
-        pulumi.set(__self__, "update", update)
-        pulumi.set(__self__, "write", write)
+        if add is not None:
+            pulumi.set(__self__, "add", add)
+        if create is not None:
+            pulumi.set(__self__, "create", create)
+        if delete is not None:
+            pulumi.set(__self__, "delete", delete)
+        if filter is not None:
+            pulumi.set(__self__, "filter", filter)
+        if list is not None:
+            pulumi.set(__self__, "list", list)
+        if process is not None:
+            pulumi.set(__self__, "process", process)
+        if read is not None:
+            pulumi.set(__self__, "read", read)
+        if tag is not None:
+            pulumi.set(__self__, "tag", tag)
+        if update is not None:
+            pulumi.set(__self__, "update", update)
+        if write is not None:
+            pulumi.set(__self__, "write", write)
 
     @_builtins.property
     @pulumi.getter
-    def add(self) -> _builtins.bool:
+    def add(self) -> Optional[_builtins.bool]:
         """
         Should Add permissions be enabled for this SAS?
         """
@@ -3563,7 +3672,7 @@ class GetAccountSASPermissionsResult(dict):
 
     @_builtins.property
     @pulumi.getter
-    def create(self) -> _builtins.bool:
+    def create(self) -> Optional[_builtins.bool]:
         """
         Should Create permissions be enabled for this SAS?
         """
@@ -3571,7 +3680,7 @@ class GetAccountSASPermissionsResult(dict):
 
     @_builtins.property
     @pulumi.getter
-    def delete(self) -> _builtins.bool:
+    def delete(self) -> Optional[_builtins.bool]:
         """
         Should Delete permissions be enabled for this SAS?
         """
@@ -3579,18 +3688,15 @@ class GetAccountSASPermissionsResult(dict):
 
     @_builtins.property
     @pulumi.getter
-    def filter(self) -> _builtins.bool:
+    def filter(self) -> Optional[_builtins.bool]:
         """
         Should Filter by Index Tags permissions be enabled for this SAS?
-
-        Refer to the [SAS creation reference from Azure](https://docs.microsoft.com/rest/api/storageservices/constructing-an-account-sas)
-        for additional details on the fields above.
         """
         return pulumi.get(self, "filter")
 
     @_builtins.property
     @pulumi.getter
-    def list(self) -> _builtins.bool:
+    def list(self) -> Optional[_builtins.bool]:
         """
         Should List permissions be enabled for this SAS?
         """
@@ -3598,7 +3704,7 @@ class GetAccountSASPermissionsResult(dict):
 
     @_builtins.property
     @pulumi.getter
-    def process(self) -> _builtins.bool:
+    def process(self) -> Optional[_builtins.bool]:
         """
         Should Process permissions be enabled for this SAS?
         """
@@ -3606,7 +3712,7 @@ class GetAccountSASPermissionsResult(dict):
 
     @_builtins.property
     @pulumi.getter
-    def read(self) -> _builtins.bool:
+    def read(self) -> Optional[_builtins.bool]:
         """
         Should Read permissions be enabled for this SAS?
         """
@@ -3614,7 +3720,7 @@ class GetAccountSASPermissionsResult(dict):
 
     @_builtins.property
     @pulumi.getter
-    def tag(self) -> _builtins.bool:
+    def tag(self) -> Optional[_builtins.bool]:
         """
         Should Get / Set Index Tags permissions be enabled for this SAS?
         """
@@ -3622,7 +3728,7 @@ class GetAccountSASPermissionsResult(dict):
 
     @_builtins.property
     @pulumi.getter
-    def update(self) -> _builtins.bool:
+    def update(self) -> Optional[_builtins.bool]:
         """
         Should Update permissions be enabled for this SAS?
         """
@@ -3630,9 +3736,11 @@ class GetAccountSASPermissionsResult(dict):
 
     @_builtins.property
     @pulumi.getter
-    def write(self) -> _builtins.bool:
+    def write(self) -> Optional[_builtins.bool]:
         """
         Should Write permissions be enabled for this SAS?
+
+        > **Note:** Refer to the [SAS creation reference from Azure](https://docs.microsoft.com/rest/api/storageservices/constructing-an-account-sas) for additional details on the fields above.
         """
         return pulumi.get(self, "write")
